+++ packages/builders/src/components/Component.ts
@@ -1,15 +1,22 @@
 import type { JSONEncodable } from '@discordjs/util';
 import type {
 	APIActionRowComponent,
-	APIActionRowComponentTypes,
+	APIComponentInActionRow,
 	APIBaseComponent,
 	ComponentType,
+	APIMessageComponent,
+	APIModalComponent,
 } from 'discord-api-types/v10';
+import { idValidator } from './Assertions';
 
 /**
  * Any action row component data represented as an object.
  */
-export type AnyAPIActionRowComponent = APIActionRowComponent<APIActionRowComponentTypes> | APIActionRowComponentTypes;
+export type AnyAPIActionRowComponent =
+	| APIActionRowComponent<APIComponentInActionRow>
+	| APIComponentInActionRow
+	| APIMessageComponent
+	| APIModalComponent;
 
 /**
  * The base component builder that contains common symbols for all sorts of components.
@@ -41,5 +48,23 @@ export abstract class ComponentBuilder<
 	 */
 	public constructor(data: Partial<DataType>) {
 		this.data = data;
+	}
+
+	/**
+	 * Sets the id (not the custom id) for this component.
+	 *
+	 * @param id - The id for this component
+	 */
+	public setId(id: number) {
+		this.data.id = idValidator.parse(id);
+		return this;
+	}
+
+	/**
+	 * Clears the id of this component, defaulting to a default incremented id.
+	 */
+	public clearId() {
+		this.data.id = undefined;
+		return this;
 	}
 }