+++ packages/builders/src/index.ts
@@ -34,6 +34,22 @@ export {
 export * from './components/selectMenu/StringSelectMenuOption.js';
 export * from './components/selectMenu/UserSelectMenu.js';
 
+export * from './components/fileUpload/FileUpload.js';
+export * as FileUploadAssertions from './components/fileUpload/Assertions.js';
+
+export * from './components/label/Label.js';
+export * as LabelAssertions from './components/label/Assertions.js';
+
+export * as ComponentsV2Assertions from './components/v2/Assertions.js';
+export * from './components/v2/Container.js';
+export * from './components/v2/File.js';
+export * from './components/v2/MediaGallery.js';
+export * from './components/v2/MediaGalleryItem.js';
+export * from './components/v2/Section.js';
+export * from './components/v2/Separator.js';
+export * from './components/v2/TextDisplay.js';
+export * from './components/v2/Thumbnail.js';
+
 export * as SlashCommandAssertions from './interactions/slashCommands/Assertions.js';
 export * from './interactions/slashCommands/SlashCommandBuilder.js';
 export * from './interactions/slashCommands/SlashCommandSubcommands.js';